simplify test with test fixtures
